+++ lib/widgets/list_view/transactions/list_view_transactions.dart
@@ -2,10 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:money/helpers/list_helper.dart';
 import 'package:money/storage/data/data.dart';
 import 'package:money/models/money_objects/transactions/transaction.dart';
-import 'package:money/widgets/confirmation_dialog.dart';
-import 'package:money/widgets/dialog.dart';
-import 'package:money/widgets/dialog_button.dart';
-
+import 'package:money/views/view_transactions/dialog_edit_transaction.dart';
 import 'package:money/widgets/list_view/list_view.dart';
 
 class ListViewTransactions extends StatefulWidget {
@@ -125,52 +122,4 @@ List<Transaction> getFilteredTransactions(final FilterFunction filter) {
     transaction.balance.value = runningBalance;
   }
   return list;
-}
-
-void showTransactionAndActions(final BuildContext context, final Transaction instance) {
-  final List<Field<Transaction, dynamic>> fields = getFieldsForClass<Transaction>()
-      .where((final Field<Transaction, dynamic> item) => item.useAsDetailPanels)
-      .toList();
-
-  final Fields<Transaction> detailPanelFields = Fields<Transaction>(definitions: fields);
-
-  Widget dialogContent = SingleChildScrollView(
-    key: Key(instance.uniqueId.toString()),
-    child: Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 8.0),
-      child: Column(
-        children: detailPanelFields.getCellsForDetailsPanel(instance),
-      ),
-    ),
-  );
-
-  myShowDialog(
-    context: context,
-    title: 'Transaction',
-    child: dialogContent,
-    actionButtons: [
-      DialogActionButton(
-        text: 'Delete',
-        onPressed: () {
-          showDialog(
-            context: context,
-            builder: (final BuildContext context) {
-              return Center(
-                child: DeleteConfirmationDialog(
-                  title: 'Delete',
-                  question: 'Are you sure you want to delete this?',
-                  content: Column(
-                    children: detailPanelFields.getListOfFieldNameAndValuePairAsWidget(instance),
-                  ),
-                  onConfirm: () {
-                    Data().transactions.deleteItem(instance);
-                  },
-                ),
-              );
-            },
-          );
-        },
-      ),
-    ],
-  );
 }